+++ src/WebSocketConnection.h
@@ -40,7 +40,7 @@ namespace Huobi {
         void reConnect(int delayInSecond);
         void close();
         void onOpen(lws* wsi);
-        void onMessage(const char* message);
+        void onMessage( char* message,size_t len);
         void send(const std::string& message);
         bool flushSendBuffer(lws* ws);
         long getLastReceivedTime();
@@ -54,9 +54,11 @@ namespace Huobi {
 
     private:
         std::string createSignature();
+        std::string createV2Signature();
         void onReceive(JsonWrapper& json);
         void processPingOnTradingLine(JsonWrapper& json);
         void processPingOnMarketLine(JsonWrapper& json);
+        void processV2Ping(JsonWrapper data);
     private:
         std::string apiKey;
         std::string secretKey;
@@ -70,8 +72,8 @@ namespace Huobi {
         long lastReceivedTime = 0;
         int delayInSecond = 0;
         std::string host;
-        std::string subscriptionMarketUrl = "wss://api.huobi.pro/ws";
-        std::string subscriptionTradingUrl = "wss://api.huobi.pro/ws/v1";
+        std::string subscriptionMarketUrl = "wss://api.huobi.so/ws";
+        std::string subscriptionTradingUrl = "wss://api.huobi.so/ws/v1";
         int connectionId;
         
         static int connectionCounter;